pyproject.toml
[project]
name = "freecodecamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "google-genai"
]
